transfer from dashboard to products
