+++ client/src/App.jsx
@@ -13,7 +13,7 @@ import { currentUser } from './functions/auth'
 import AdminDashboard from './pages/admin/AdminDashboard'
 import { CategoryCreate, CategoryUpdate } from './pages/admin/category/index'
 import { SubCreate, SubUpdate } from './pages/admin/sub/index'
-import {CreateProduct} from './pages/admin/product/index'
+import {CreateProduct, AllProducts} from './pages/admin/product/index'
 
 const App = () => {
 
@@ -61,6 +61,7 @@ const App = () => {
         <AdminRoute exact path="/admin/sub" component={SubCreate} />
         <AdminRoute exact path="/admin/sub/:slug" component={SubUpdate} />
         <AdminRoute exact path="/admin/product" component={CreateProduct} />
+        <AdminRoute exact path="/admin/products" component={AllProducts} />
       </Switch>
     </>
   )
