+++ client/src/components/cards/AdminProductCard.jsx
@@ -11,7 +11,7 @@ const AdminProductCard = ({ product }) => {
         <img
           src={images && images.length ? images[0].url : ""}
           style={{ height: '150px', objectFit: 'cover' }}
-          className="m-2"
+          className="p-1"
         />
       }
     >
