+++ client/src/pages/admin/product/index.js
@@ -1,1 +1,2 @@
-export {default as CreateProduct} from './CreateProduct'
+export { default as CreateProduct } from './CreateProduct'
+export {default as AllProducts} from './AllProducts'
